// Interfaces/userCredentials.tsx
interface userCredentials {
  address1: string;
  address2: string;
  businessDescription: string;
  businessName: string;
  state: string;
  email: string;
  firstName: string;
  activeTier: string;
  lastName: string;
  phoneNumber1: number;
  phoneNumber2: number;
  zipCode: number;
  userName: string;
  storeID: string;
  timeStamp: any;
}

export default userCredentials;